requirements.txt
# Core web and server components
dash==3.2.0
dash-table
gunicorn

# Data processing and scientific libraries
netCDF4==1.7.2
numpy==2.2.5
pandas==2.2.3
scipy==1.15.3
scikit-learn==1.7.1
pybaselines==1.2.0

# Plotting
matplotlib==3.10.1
plotly==6.2.0

# Report generation (DOCX)
python-docx
docxcompose
docxtpl==0.20.0
lxml==5.4.0

# Utilities
Deprecated==1.2.18
